+++ lib/cli.js
@@ -259,7 +259,7 @@ exports.run = async () => { // eslint-disable-line complexity
 
 	const ciParallelVars = require('ci-parallel-vars');
 	const Api = require('./api');
-	const Reporter = require('./reporters/default');
+	const DefaultReporter = require('./reporters/default');
 	const TapReporter = require('./reporters/tap');
 	const Watcher = require('./watcher');
 	const normalizeExtensions = require('./extensions');
@@ -383,30 +383,15 @@ exports.run = async () => { // eslint-disable-line complexity
 		workerArgv: argv['--']
 	});
 
-	let reporter;
-	if (combined.tap && !combined.watch && debug === null) {
-		reporter = new TapReporter({
-			projectDir,
-			reportStream: process.stdout,
-			stdStream: process.stderr
-		});
-	} else if (debug !== null || combined.verbose || isCi || !process.stdout.isTTY) {
-		reporter = new VerboseReporter({
-			projectDir,
-			reportStream: process.stdout,
-			stdStream: process.stderr,
-			watching: combined.watch,
-			mode: 'verbose'
-		});
-	} else {
-		reporter = new MiniReporter({
-			projectDir,
-			reportStream: process.stdout,
-			stdStream: process.stderr,
-			watching: combined.watch,
-			mode: 'verbose'
-		});
-	}
+	const Reporter = (combined.tap && !combined.watch && debug === null) ? TapReporter : DefaultReporter;
+
+	const reporter = new Reporter({
+		projectDir,
+		reportStream: process.stdout,
+		stdStream: process.stderr,
+		watching: combined.watch,
+		mode: (debug !== null || combined.verbose || isCi || !process.stdout.isTTY) ? 'verbose' : 'mini'
+	});
 
 	api.on('run', plan => {
 		reporter.startRun(plan);
